correct files/folders structure
