+++ schemas/contacts.js
@@ -1,4 +1,4 @@
-const mongoose = require("../db");
+const mongoose = require("mongoose");
 const Schema = mongoose.Schema;
 const contacts = new Schema({
   name: { type: String, required: [true, "Set name for contact"] },
